requirements.txt
aiohttp==3.5.4
discord.py==1.3.3
geocoder==1.38.1
pyjokes==0.6.0
lyricsgenius==1.8.5